backend/requirements.txt
Flask
flask_sqlalchemy
Authlib
SQLAlchemy 
flask-socketio
Flask-Session
flask-cors
requests
dotenv
pandas
pathlib
apscheduler
urllib3
gevent
gevent-websocket
gunicorn
pytz
openpyxl
qdrant-client
PyPDF2
python-docx
psycopg2-binary
pycryptodome
pytesseract
Pillow
PyMuPDF
blinker